// libs/plugins/test-runner/src/go-judge-api.d.ts
export namespace GoJudge {
  type RunResult = Result[];
// ...
  interface Result {
    status: Status;
    error?: string; // potential system error message
    exitStatus: number;
    time: number; // ns (cgroup recorded time)
    memory: number; // byte
    runTime: number; // ns (wall clock time)
    procPeak?: number; // peak number of process (cgroup v2, kernel >= 6.1)
    // copyFile name -> content
    files?: { [name: string]: string };
    // copyFileCached name -> fileId
    fileIds?: { [name: string]: string };
    // fileError contains detailed file errors
    fileError?: FileError[];
  }
// ...
}